backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.35
aiosqlite==0.20.0
langchain==0.3.7
langchain-openai==0.2.6
langchain-core==0.3.18
langchain-community==0.3.5
openai==1.54.4
aiofiles==0.8.0
python-docx==1.1.2
reportlab==4.2.5
matplotlib==3.9.2
openpyxl==3.1.5
pandas==2.2.3
pypdf==4.2.0
